# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: KYAgent API
  version: 1.0.0
  description: |
    Know Your Agent — identity verification and authorization for autonomous AI agents.
    Normative companions: ARCHITECTURE.md, docs/contracts/types.ts, error-model.md,
    crypto-and-signing.md. All request bodies are strict (additionalProperties: false).
    Authentication: `Authorization: Bearer kya_<keyId>_<secret>`. RBAC per ARCHITECTURE.md §5.
    Cross-tenant access returns 404.
servers:
  - url: http://localhost:8080
security:
  - apiKey: []
...
paths:
...
  /v1/businesses:
    post:
      tags: [businesses]
      operationId: createBusiness
      x-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBusinessRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Business' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
    get:
      tags: [businesses]
      operationId: listBusinesses
      x-roles: [admin]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Page of businesses
          content:
            application/json:
              schema:
                type: object
                required: [data, nextCursor]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Business' } }
                  nextCursor: { type: [string, 'null'] }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/businesses/{id}:
    get:
      tags: [businesses]
      operationId: getBusiness
      x-roles: [admin, business(own)]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Business
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Business' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }

  # ------------------------------------------------------------------ operators
  /v1/operators:
    post:
      tags: [operators]
      operationId: createOperator
      x-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOperatorRequest' }
      responses:
        '201':
          description: Created (status pending)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operator' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
    get:
      tags: [operators]
      operationId: listOperators
      x-roles: [admin]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/OperatorStatus' }
      responses:
        '200':
          description: Page of operators
          content:
            application/json:
              schema:
                type: object
                required: [data, nextCursor]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Operator' } }
                  nextCursor: { type: [string, 'null'] }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/operators/{id}:
    get:
      tags: [operators]
      operationId: getOperator
      x-roles: [admin, operator(own), business]
      description: Admin and the owning operator get `Operator`; businesses get `OperatorPublicProfile`.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Operator
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Operator'
                  - $ref: '#/components/schemas/OperatorPublicProfile'
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/operators/{id}/verification:
    post:
      tags: [operators]
      operationId: verifyOperator
      x-roles: [admin]
      description: |
        Runs mock KYC + sanctions screening (data-schema.md "Mock KYC/sanctions").
        Allowed from `pending` or `rejected`; otherwise 409 INVALID_STATE.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Operator with verification result (status verified or rejected)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operator' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /v1/operators/{id}/suspend:
    post:
      tags: [operators]
      operationId: suspendOperator
      x-roles: [admin]
      description: Suspends a verified operator. All its agents DENY with OPERATOR_SUSPENDED; its API keys stop authenticating.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReasonRequest' }
      responses:
        '200':
          description: Suspended operator
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operator' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }

  # ------------------------------------------------------------------ api keys
  /v1/api-keys:
    post:
      tags: [api-keys]
      operationId: createApiKey
      x-roles: [admin]
      description: |
        `ownerId` required (and must exist) for role operator/business; must be null/absent for admin.
        The plaintext `secret` is returned only in this response.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateApiKeyRequest' }
      responses:
        '201':
          description: Created
          headers:
            Cache-Control: { schema: { type: string, const: no-store } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreateApiKeyResponse' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
    get:
      tags: [api-keys]
      operationId: listApiKeys
      x-roles: [admin]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: ownerId
          in: query
          schema: { type: string, maxLength: 64 }
      responses:
        '200':
          description: Page of API keys (never includes secrets or hashes)
          content:
            application/json:
              schema:
                type: object
                required: [data, nextCursor]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/ApiKey' } }
                  nextCursor: { type: [string, 'null'] }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/api-keys/{id}/revoke:
    post:
      tags: [api-keys]
      operationId: revokeApiKey
      x-roles: [admin]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Revoked key
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiKey' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }

  # ------------------------------------------------------------------ agents
  /v1/agents:
    post:
      tags: [agents]
      operationId: registerAgent
      x-roles: [operator]
      description: |
        Registers an agent bound to the calling operator. Operator must be `verified`
        (else 409 OPERATOR_NOT_VERIFIED). `proofOfPossession` per crypto-and-signing.md §2.
        Duplicate public key => 409 CONFLICT.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAgentRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
    get:
      tags: [agents]
      operationId: listAgents
      x-roles: [admin, operator(own)]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: operatorId
          in: query
          description: Admin only filter; ignored (forced to self) for operators.
          schema: { type: string, maxLength: 64 }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/AgentStatus' }
      responses:
        '200':
          description: Page of agents
          content:
            application/json:
              schema:
                type: object
                required: [data, nextCursor]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Agent' } }
                  nextCursor: { type: [string, 'null'] }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/agents/{id}:
    get:
      tags: [agents]
      operationId: getAgent
      x-roles: [admin, operator(own), business]
      description: Identity lookup. Businesses may read any agent (public identity).
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Agent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/agents/{id}/suspend:
    post:
      tags: [agents]
      operationId: suspendAgent
      x-roles: [admin, operator(own)]
      description: active -> suspended. Verifications DENY with AGENT_SUSPENDED.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReasonRequest' }
      responses:
        '200':
          description: Agent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /v1/agents/{id}/reactivate:
    post:
      tags: [agents]
      operationId: reactivateAgent
      x-roles: [admin, operator(own)]
      description: suspended -> active. Not allowed from revoked (409 INVALID_STATE).
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Agent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /v1/agents/{id}/revoke:
    post:
      tags: [agents]
      operationId: revokeAgent
      x-roles: [admin, operator(own)]
      description: Permanent. Effective immediately for all subsequent verifications (AGENT_REVOKED).
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReasonRequest' }
      responses:
        '200':
          description: Revoked agent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /v1/agents/{id}/credentials:
    post:
      tags: [credentials]
      operationId: issueCredential
      x-roles: [operator(own)]
      description: Issues a signed, expiring, scoped, key-bound credential for a grant (crypto-and-signing.md §4).
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IssueCredentialRequest' }
      responses:
        '201':
          description: Issued
          headers:
            Cache-Control: { schema: { type: string, const: no-store } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IssueCredentialResponse' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }

  # ------------------------------------------------------------------ grants
  /v1/grants:
    post:
      tags: [grants]
      operationId: createGrant
      x-roles: [business]
      description: Calling business authorizes an active agent for actions until expiresAt.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateGrantRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Grant' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
    get:
      tags: [grants]
      operationId: listGrants
      x-roles: [admin, business(own), operator(own agents)]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: agentId
          in: query
          schema: { type: string, maxLength: 64 }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/GrantStatus' }
      responses:
        '200':
          description: Page of grants
          content:
            application/json:
              schema:
                type: object
                required: [data, nextCursor]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Grant' } }
                  nextCursor: { type: [string, 'null'] }
        '401': { $ref: '#/components/responses/Error' }
  /v1/grants/{id}:
    get:
      tags: [grants]
      operationId: getGrant
      x-roles: [admin, business(own), operator(own agents)]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Grant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Grant' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/grants/{id}/revoke:
    post:
      tags: [grants]
      operationId: revokeGrant
      x-roles: [admin, business(own)]
      description: Permanent. Credentials derived from it DENY with GRANT_REVOKED.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReasonRequest' }
      responses:
        '200':
          description: Revoked grant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Grant' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }

  # ------------------------------------------------------------------ credentials
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: 'kya_<keyId>_<secret>'
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, pattern: '^[a-z]{2,3}_[0-9A-HJKMNP-TV-Z]{26}$' }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string, maxLength: 200 }
  responses:
    Error:
      description: Error envelope (error-model.md)
      headers:
        X-Request-Id: { schema: { type: string } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
  schemas:
...
    Business:
      type: object
      required: [id, name, status, createdAt]
      properties:
        id: { $ref: '#/components/schemas/ResourceId' }
        name: { type: string }
        status: { type: string, enum: [active, suspended] }
        createdAt: { $ref: '#/components/schemas/IsoDateTime' }
    CreateBusinessRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }

    # ---------- operators
    OperatorStatus: { type: string, enum: [pending, verified, rejected, suspended] }
...